requirements.txt
streamlitgit 
tensorflow==2.20.0
Pillow
numpy
gunicorn